requirements.txt
pytest
pytest-html
pytest-playwright
google-generativeai
python-dotenv
pydantic